requirements.txt
gevent-websocket
flask
stomp.py